Prevent an illegal use of a generic from crashing the compiler (#3277)
