+++ test/src/e2e_vm_tests/test_programs/should_fail/abi_method_signature_mismatch/test.toml
@@ -1,13 +1,13 @@
 category = "fail"
 
 # check: fn foo(s: str[7]) -> str[7] {
-# nextln: $()Expected: u64
-# nextln: $()found:    str[7]. The definition of this function must match the one in the ABI "MyContract" declaration.
+# check: $()Expected: u64
+# check: $()found:    str[7]. The definition of this function must match the one in the ABI "MyContract" declaration.
 
 # check: fn bar() -> u64 {
-# nextln: $()Expected: u32
-# nextln: $()found:    u64. The definition of this function must match the one in the ABI "MyContract" declaration.
+# check: $()Expected: u32
+# check: $()found:    u64. The definition of this function must match the one in the ABI "MyContract" declaration.
 
 # check: fn baz() {
-# nextln: $()Expected: u64
-# nextln: $()found:    (). The definition of this function must match the one in the ABI "MyContract" declaration.
+# check: $()Expected: u64
+# check: $()found:    (). The definition of this function must match the one in the ABI "MyContract" declaration.
