+++ test/src/e2e_vm_tests/test_programs/should_fail/abi_pure_calls_impure/test.toml
@@ -1,4 +1,4 @@
 category = "fail"
 
 # check: f()
-# nextln: $()Storage attribute access mismatch. Try giving the surrounding function more access by adding "#[storage(read)]" to the function declaration.
+# check: $()Storage attribute access mismatch. Try giving the surrounding function more access by adding "#[storage(read)]" to the function declaration.
