+++ test/src/e2e_vm_tests/test_programs/should_fail/contract_pure_calls_impure/test.toml
@@ -1,4 +1,4 @@
 category = "fail"
 
 # check: foo();
-# nextln: $()Storage attribute access mismatch. Try giving the surrounding function more access by adding "#[storage(read, write)]" to the function declaration.
+# check: $()Storage attribute access mismatch. Try giving the surrounding function more access by adding "#[storage(read, write)]" to the function declaration.
