+++ test/src/e2e_vm_tests/test_programs/should_fail/illegal_use_of_generic/Forc.lock
@@ -1,0 +1,3 @@
+[[package]]
+name = 'illegal_use_of_generic'
+source = 'member'
