+++ test/src/e2e_vm_tests/test_programs/should_fail/illegal_use_of_generic/Forc.toml
@@ -1,0 +1,6 @@
+[project]
+name = "illegal_use_of_generic"
+authors = ["Fuel Labs <contact@fuel.sh>"]
+entry = "main.sw"
+license = "Apache-2.0"
+implicit-std = false
