+++ test/src/e2e_vm_tests/test_programs/should_fail/illegal_use_of_generic/src/main.sw
@@ -1,0 +1,20 @@
+script;
+
+// We want this to error without crashing the compiler.
+fn test_function<T>(value: T) {
+    T::new();
+}
+
+struct Data<T> {
+    value: T
+}
+
+impl<T> Data<T> {
+    fn test_function(self) {
+        T::new();
+    }
+}
+
+fn main() {
+
+}
