+++ test/src/e2e_vm_tests/test_programs/should_fail/impure_abi_read_calls_impure_write/test.toml
@@ -1,5 +1,5 @@
 category = "fail"
 
 # check: fn test_function() -> bool {
-# nextln: f()
-# nextln: $()Storage attribute access mismatch. Try giving the surrounding function more access by adding "#[storage(read, write)]" to the function declaration.
+# check: f()
+# check: $()Storage attribute access mismatch. Try giving the surrounding function more access by adding "#[storage(read, write)]" to the function declaration.
