+++ test/src/e2e_vm_tests/test_programs/should_fail/trait_method_signature_mismatch/test.toml
@@ -1,16 +1,16 @@
 category = "fail"
 
 # check: fn foo(s: str[7]) -> str[7] {
-# nextln: $()Expected: u64
-# nextln: $()found:    str[7]. The definition of this function must match the one in the trait "Foo" declaration.
+# check: $()Expected: u64
+# check: $()found:    str[7]. The definition of this function must match the one in the trait "Foo" declaration.
 
 # check: fn bar(ref mut variable: u64) -> bool {
-# nextln: $()Parameter mutability mismatch between the trait function declaration and its implementation.
+# check: $()Parameter mutability mismatch between the trait function declaration and its implementation.
 
 # check: fn baz() -> u64 {
-# nextln: $()Expected: u32
-# nextln: $()found:    u64. The definition of this function must match the one in the trait "Foo" declaration.
+# check: $()Expected: u32
+# check: $()found:    u64. The definition of this function must match the one in the trait "Foo" declaration.
 
 # check: fn quux() {
-# nextln: $()Expected: u64
-# nextln: $()found:    (). The definition of this function must match the one in the trait "Foo" declaration.
+# check: $()Expected: u64
+# check: $()found:    (). The definition of this function must match the one in the trait "Foo" declaration.
