+++ test/src/e2e_vm_tests/test_programs/should_fail/where_clause_functions/test.toml
@@ -1,7 +1,7 @@
 category = "fail"
 
 # check: $()fn add_points<T>(a: MyPoint<T>, b: MyPoint<T>) -> MyPoint<T> where T: MyAdd {
-# nextln: $()"where" clauses are not yet supported
+# check: $()"where" clauses are not yet supported
 
 # check: $()let baz = add_points(foo, bar);
-# nextln: $()Could not find symbol "add_points" in this scope.
+# check: $()Could not find symbol "add_points" in this scope.
