+++ test/src/e2e_vm_tests/test_programs/should_fail/where_clause_impls/test.toml
@@ -1,7 +1,7 @@
 category = "fail"
 
 # check: $()impl<T> MyAdd for MyPoint<T> where T: MyAdd {
-# nextln: $()"where" clauses are not yet supported
+# check: $()"where" clauses are not yet supported
 
 # check: $()let baz = foo.my_add(bar);
-# nextln: $()No method named "my_add" found for type "MyPoint<u32>".
+# check: $()No method named "my_add" found for type "MyPoint<u32>".
